requirements.txt
django==2.2
psycopg2==2.8.6
djangorestframework==3.13.1

